Reorganize code into standard python package structure
- removed example arguments
- switch imports to be relative (dot imports)
