+++ examples/Macau_example.args
